# packages/api-client/openapi.json
{"openapi":"3.1.0","info":{"title":"RunbookOS Control Plane","version":"0.4.0"},"servers":[{"url":"http://localhost:8080","description":"Local development"}],"paths":{"/api/webhooks/{organizationId}/{source}":{"post":{"tags":["webhook-controller"],"operationId":"receive","parameters":[{"name":"organizationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"source","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Delivery-ID","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Webhook-Nonce","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Webhook-Timestamp","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Webhook-Signature","in":"header","required":false,"schema":{"type":"string"}},{"name":"X-Hub-Signature-256","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"byte"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WebhookResponse"}}}}}}},"/api/organizations":{"post":{"tags":["organization-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrganizationRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationResponse"}}}}}}},"/api/organizations/{organizationId}/switch":{"post":{"tags":["organization-controller"],"operationId":"switchOrganization","parameters":[{"name":"organizationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SwitchOrganizationResponse"}}}}}}},"/api/internal/workflows/callbacks":{"post":{"tags":["workflow-callback-controller"],"operationId":"callback","parameters":[{"name":"X-Execution-Token","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Internal-Timestamp","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Internal-Nonce","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Internal-Signature","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"byte"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/incidents/{id}/transitions":{"post":{"tags":["incident-controller"],"operationId":"transition","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransitionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentView"}}}}}}},"/api/incidents/{id}/comments":{"post":{"tags":["incident-controller"],"operationId":"comment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentView"}}}}}}},"/api/incidents/{id}/assignments":{"post":{"tags":["incident-controller"],"operationId":"assign","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentView"}}}}}}},"/api/incidents/demo":{"post":{"tags":["incident-controller"],"operationId":"demo","responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentView"}}}}}}},"/api/executions":{"post":{"tags":["execution-controller"],"operationId":"create_1","parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExecutionRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExecutionView"}}}}}}},"/api/executions/{id}/steps/{actionId}/retry":{"post":{"tags":["execution-controller"],"operationId":"retryStep","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"actionId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StepView"}}}}}}},"/api/executions/{id}/retry":{"post":{"tags":["execution-controller"],"operationId":"retry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExecutionView"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/incidents":{"get":{"tags":["incident-controller"],"operationId":"search","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["DETECTED","TRIAGED","INVESTIGATING","AWAITING_APPROVAL","MITIGATING","MONITORING","RESOLVED","CLOSED","CANCELLED"]}},{"name":"severity","in":"query","required":false,"schema":{"type":"string","enum":["SEV1","SEV2","SEV3","SEV4"]}},{"name":"serviceName","in":"query","required":false,"schema":{"type":"string"}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"cursor","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":25}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentPage"}}}}}}},"/api/incidents/{id}":{"get":{"tags":["incident-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentDetail"}}}}}}},"/api/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/executions/{id}":{"get":{"tags":["execution-controller"],"operationId":"timeline","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExecutionTimeline"}}}}}}},"/api/executions/{id}/events":{"get":{"tags":["execution-controller"],"operationId":"events","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"current","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentUserResponse"}}}}}}}},"components":{"schemas":{"WebhookResponse":{"type":"object","properties":{"incidentId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["DETECTED","TRIAGED","INVESTIGATING","AWAITING_APPROVAL","MITIGATING","MONITORING","RESOLVED","CLOSED","CANCELLED"]},"signalCount":{"type":"integer","format":"int32"}}},"CreateOrganizationRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":120,"minLength":0},"slug":{"type":"string","maxLength":64,"minLength":0},"demoMode":{"type":"boolean"}},"required":["name"]},"OrganizationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"demoMode":{"type":"boolean"}}},"SwitchOrganizationResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"organizationId":{"type":"string","format":"uuid"}}},"TransitionRequest":{"type":"object","properties":{"status":{"type":"string","enum":["DETECTED","TRIAGED","INVESTIGATING","AWAITING_APPROVAL","MITIGATING","MONITORING","RESOLVED","CLOSED","CANCELLED"]}}},"IncidentView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"summary":{"type":"string"},"status":{"type":"string","enum":["DETECTED","TRIAGED","INVESTIGATING","AWAITING_APPROVAL","MITIGATING","MONITORING","RESOLVED","CLOSED","CANCELLED"]},"severity":{"type":"string","enum":["SEV1","SEV2","SEV3","SEV4"]},"priority":{"type":"integer","format":"int32"},"affectedService":{"type":"string"},"assigneeUserId":{"type":"string","format":"uuid"},"signalCount":{"type":"integer","format":"int32"},"firstDetectedAt":{"type":"string","format":"date-time"},"lastSignalAt":{"type":"string","format":"date-time"}}},"CommentRequest":{"type":"object","properties":{"body":{"type":"string","minLength":1}},"required":["body"]},"CommentView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"authorUserId":{"type":"string","format":"uuid"},"body":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"AssignRequest":{"type":"object","properties":{"assigneeUserId":{"type":"string","format":"uuid"}}},"CreateExecutionRequest":{"type":"object","properties":{"incidentId":{"type":"string","format":"uuid"},"workflowKey":{"type":"string","minLength":1}},"required":["workflowKey"]},"ExecutionView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"incidentId":{"type":"string","format":"uuid"},"workflowKey":{"type":"string"},"workflowVersion":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["QUEUED","RUNNING","PAUSED_FOR_APPROVAL","SUCCEEDED","PARTIALLY_SUCCEEDED","FAILED","TIMED_OUT","CANCELLED"]},"allowedActionIds":{"type":"array","items":{"type":"string"}},"timeoutAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"StepView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"actionId":{"type":"string"},"name":{"type":"string"},"sequence":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["QUEUED","RUNNING","WAITING","APPROVED","DENIED","SUCCEEDED","FAILED","SKIPPED"]},"attempt":{"type":"integer","format":"int32"},"startedAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"},"errorCode":{"type":"string"},"errorMessage":{"type":"string"},"output":{"type":"object","additionalProperties":{}}}},"SignupRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":128,"minLength":12},"displayName":{"type":"string","maxLength":120,"minLength":0}},"required":["displayName","email"]},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserView"}}},"UserView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"IncidentPage":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/IncidentView"}},"nextCursor":{"type":"string"},"hasMore":{"type":"boolean"}}},"EvidenceView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"kind":{"type":"string"},"title":{"type":"string"},"content":{"type":"object","additionalProperties":{}},"collectedAt":{"type":"string","format":"date-time"}}},"IncidentDetail":{"type":"object","properties":{"incident":{"$ref":"#/components/schemas/IncidentView"},"evidence":{"type":"array","items":{"$ref":"#/components/schemas/EvidenceView"}},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentView"}}}},"ComponentHealth":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"error":{"type":"string"}}},"HealthResponse":{"type":"object","properties":{"status":{"type":"string"},"timestamp":{"type":"string"},"uptime":{"type":"string"},"components":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/ComponentHealth"}}}},"EventView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"stepExecutionId":{"type":"string","format":"uuid"},"type":{"type":"string"},"status":{"type":"string"},"message":{"type":"string"},"details":{"type":"object","additionalProperties":{}},"occurredAt":{"type":"string","format":"date-time"}}},"ExecutionTimeline":{"type":"object","properties":{"execution":{"$ref":"#/components/schemas/ExecutionView"},"steps":{"type":"array","items":{"$ref":"#/components/schemas/StepView"}},"events":{"type":"array","items":{"$ref":"#/components/schemas/EventView"}}}},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"CurrentUserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"selectedOrganizationId":{"type":"string","format":"uuid"},"organizations":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationMembershipResponse"}}}},"OrganizationMembershipResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"demoMode":{"type":"boolean"},"role":{"type":"string"}}},"ErrorResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"},"correlationId":{"type":"string"},"validationErrors":{"type":"array","items":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"},"rejectedValue":{"type":"object"}}}}}}}}}